changed t to float

+++ money-change.cpp
@@ -1,8 +1,9 @@
 #include<iostream>
 #include<algorithm>
 using namespace std;
+
 int main() {
-	int t;
+	float t;
 	cin>>t;
 	while(t--){
 		int n;
